Move manual bindings and metadata into Swift directory
