+++ src/Swift.Bindings/tests/IntegrationTests/CMakeLists.txt
@@ -39,7 +39,14 @@ foreach(SOURCE_FILE ${SOURCES})
     get_filename_component(DOTNET_CONFIGURATION ${DOTNET_CONFIGURATION} NAME)
 
     add_custom_command(TARGET ${SOURCE_BASE_NAME} POST_BUILD
-        COMMAND ${CMAKE_SOURCE_DIR}/../../../../dotnet.sh ${CMAKE_BINARY_DIR}/../../../Swift.Bindings/${DOTNET_CONFIGURATION}/${DOTNET_VERSION}/Swift.Bindings.dll -a ${OUTPUT_DIR}/${SOURCE_BASE_NAME}.abi.json -o ${OUTPUT_DIR}
+        COMMAND ${CMAKE_SOURCE_DIR}/../../../../dotnet.sh ${CMAKE_BINARY_DIR}/../../../Swift.Bindings/${DOTNET_CONFIGURATION}/${DOTNET_VERSION}/Swift.Bindings.dll -a ${OUTPUT_DIR}/${SOURCE_BASE_NAME}.abi.json -d ${OUTPUT_DIR}/lib${SOURCE_BASE_NAME}.dylib -o ${OUTPUT_DIR}
         COMMENT "Generating ${SOURCE_BASE_NAME} bindings"
     )
+
+    add_custom_command(TARGET ${SOURCE_BASE_NAME} POST_BUILD
+        COMMAND ${CMAKE_COMMAND} -E remove_directory ${OUTPUT_DIR}/Swift
+        COMMAND ${CMAKE_COMMAND} -E remove_directory ${OUTPUT_DIR}/../Swift
+        COMMENT "Remove redundant Swift directory"
+    )
+
 endforeach()
