+++ src/Swift.Runtime/src/Swift/Runtime/ITypeMetadataCache.cs
