+++ src/Swift.Runtime/src/Swift/Runtime/KnownLibraries.cs
@@ -1,3 +1,5 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
 
 namespace Swift.Runtime;
