+++ src/Swift.Runtime/src/Swift/Runtime/SwiftHandle.cs
@@ -1,3 +1,6 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
 using System;
 using System.Collections.Generic;
