+++ src/Swift.Runtime/src/Swift/Runtime/TypeMetadata.cs
