+++ src/Swift.Runtime/src/Swift/Runtime/TypeMetadataCache.cs
