--- src/Swift.Runtime/src/TypeDatabase.xml
@@ -1,71 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<swifttypedatabase version="1.0">
-    <entities>
-        <entity managedNameSpace="System" managedTypeName="void">
-            <typedeclaration kind="typealias" name="()" module="" frozen="false" blittable ="false" />
-        </entity>
-        <entity managedNameSpace="System" managedTypeName="Byte">
-            <typedeclaration kind="struct" name="UInt8" module="Swift" metadataAccessor="$ss5UInt8VMa" frozen="true" blittable ="true" />
-        </entity>
-        <entity managedNameSpace="System" managedTypeName="SByte">
-            <typedeclaration kind="struct" name="Int8" module="Swift" metadataAccessor="$ss4Int8VMa" frozen="true" blittable ="true" />
-        </entity>
-        <entity managedNameSpace="System" managedTypeName="Int16">
-            <typedeclaration kind="struct" name="Int16" module="Swift" metadataAccessor="$ss5Int16VMa" frozen="true" blittable ="true" />
-        </entity>
-        <entity managedNameSpace="System" managedTypeName="UInt16">
-            <typedeclaration kind="struct" name="UInt16" module="Swift" metadataAccessor="$ss6UInt16VMa" frozen="true" blittable ="true" />
-        </entity>
-        <entity managedNameSpace="System" managedTypeName="Int32">
-            <typedeclaration kind="struct" name="Int32" module="Swift" metadataAccessor="$ss5Int32VMa" frozen="true" blittable ="true" />
-        </entity>
-        <entity managedNameSpace="System" managedTypeName="UInt32">
-            <typedeclaration kind="struct" name="UInt32" module="Swift" metadataAccessor="$ss6UInt32VMa" frozen="true" blittable ="true" />
-        </entity>
-        <entity managedNameSpace="System" managedTypeName="Int64">
-            <typedeclaration kind="struct" name="Int64" module="Swift" metadataAccessor="$ss5Int64VMa" frozen="true" blittable ="true" />
-        </entity>
-        <entity managedNameSpace="System" managedTypeName="UInt64">
-            <typedeclaration kind="struct" name="UInt64" module="Swift" metadataAccessor="$ss6UInt64VMa" frozen="true" blittable ="true" />
-        </entity>
-        <entity managedNameSpace="System" managedTypeName="Single">
-            <typedeclaration kind="struct" name="Float" module="Swift" metadataAccessor="$sSfMa" frozen="true" blittable ="true" />
-        </entity>
-        <entity managedNameSpace="System" managedTypeName="Double">
-            <typedeclaration kind="struct" name="Double" module="Swift" metadataAccessor="$sSdMa" frozen="true" blittable ="true" />
-        </entity>
-        <entity managedNameSpace="System" managedTypeName="IntPtr">
-            <typedeclaration kind="struct" name="Int" module="Swift" metadataAccessor="$sSiMa" frozen="true" blittable ="true" />
-        </entity>
-        <entity managedNameSpace="System" managedTypeName="UIntPtr">
-            <typedeclaration kind="struct" name="UInt" module="Swift" metadataAccessor="$sSuMa" frozen="true" blittable ="true" />
-        </entity>
-        <entity managedNameSpace="System" managedTypeName="bool">
-            <typedeclaration kind="struct" name="Bool" module="Swift" metadataAccessor="$sSbMa" frozen="true" blittable ="true" />
-        </entity>
-        <entity managedNameSpace="Swift.Runtime" managedTypeName="UnsafePointer">
-            <typedeclaration kind="struct" name="UnsafePointer" module="Swift" metadataAccessor="sSPMa" frozen="true" blittable ="true" />
-        </entity>
-        <entity managedNameSpace="Swift.Runtime" managedTypeName="UnsafeMutablePointer">
-            <typedeclaration kind="struct" name="UnsafeMutablePointer" module="Swift" metadataAccessor="sSpMa" frozen="true" blittable ="true" />
-        </entity>
-        <entity managedNameSpace="Swift.Runtime" managedTypeName="UnsafeRawPointer">
-            <typedeclaration kind="struct" name="UnsafeRawPointer" module="Swift" metadataAccessor="sSVMa" frozen="true" blittable ="true"/>
-        </entity>
-        <entity managedNameSpace="Swift.Runtime" managedTypeName="UnsafeMutableRawPointer">
-            <typedeclaration kind="struct" name="UnsafeMutableRawPointer" module="Swift" metadataAccessor="sSvMa" frozen="true" blittable ="true"/>
-        </entity>
-        <entity managedNameSpace="Swift.Runtime" managedTypeName="UnsafeBufferPointer">
-            <typedeclaration kind="struct" name="UnsafeBufferPointer" module="Swift" metadataAccessor="sSRMa" frozen="true" blittable ="true"/>
-        </entity>
-        <entity managedNameSpace="Swift.Runtime" managedTypeName="UnsafeMutableBufferPointer">
-            <typedeclaration kind="struct" name="UnsafeMutableBufferPointer" module="Swift" metadataAccessor="sSrMa" frozen="true" blittable ="true"/>
-        </entity>
-        <entity managedNameSpace="Swift.Runtime" managedTypeName="UnsafeRawBufferPointer">
-            <typedeclaration kind="struct" name="UnsafeRawBufferPointer" module="Swift" metadataAccessor="sSWMa" frozen="true" blittable ="true"/>
-        </entity>
-        <entity managedNameSpace="Swift.Runtime" managedTypeName="UnsafeMutableRawBufferPointer">
-            <typedeclaration kind="struct" name="UnsafeMutableRawBufferPointer" module="Swift" metadataAccessor="sSwMa" frozen="true" blittable ="true"/>
-        </entity>
-    </entities>
-</swifttypedatabase>
